brique39/dags/requirements.txt
apache-airflow==2.7.1
pandas==2.0.3  # Version plus stable avec wheels pré-compilés
scikit-learn==1.3.2
lightgbm==4.1.0
joblib==1.3.2
numpy==1.24.3
pyarrow==13.0.0